legacy/frontend/requirement.txt
# python==3.10.14
Flask==3.0.3
requests==2.28.1
pillow==10.4.0
gevent==24.2.1
gunicorn==23.0.0
influxdb==5.3.2
python-dotenv